requirements.txt
pillow
protobuf
addict
future
lmdb
numpy
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
tqdm
yapf
lpips
insightface
basicsr
facexlib  
onnx==1.16.0 
onnxruntime==1.17.3